requirements.txt
streamlit
google-generativeai
chromadb
pandas